Write event in sync
Problem with writing event async is that it is not displayed after submission until the page is refreshed.
This leads to confusing UI.

However, the async way should be the default way for writing to Influx, hence the API for write is not changed, only additional method `write_record_sync` is added.